feat: ajout d'une navbar fonctionnelle
