+++ .fvmrc
@@ -1,0 +1,3 @@
+{
+  "flutter": "3.35.3"
+}
