+++ lib/features/home/presentation/providers/home_provider.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart';
 import '../../domain/entities/home_data.dart';
 import '../../domain/repositories/home_repository.dart';
