+++ lib/features/home/presentation/widgets/recent_activity.dart
@@ -95,7 +95,7 @@ class _ActivityItemCard extends StatelessWidget {
           Container(
             padding: const EdgeInsets.all(8),
             decoration: BoxDecoration(
-              color: activity.color.withOpacity(0.1),
+              color: activity.color.withValues(alpha: 0.1),
               shape: BoxShape.circle,
             ),
             child: Icon(activity.icon, color: activity.color, size: 20),
